+++ lib/LocalizedObject.php
@@ -8,8 +8,6 @@ use ICanBoogie\Accessor\AccessorTrait;
  * Representation of a localized object.
  *
  * @template T of object
- *
- * @property-read Locale $locale The locale used by the formatter.
  */
 abstract class LocalizedObject
 {
@@ -20,27 +18,11 @@ abstract class LocalizedObject
 	use AccessorTrait;
 
 	/**
-	 * Creates a localized instance from the specified source and location.
-	 *
-	 * @param T $source
-	 * @param Locale $locale
-	 * @param array<string, mixed> $options
-	 *
-	 * @return static<T>
-	 */
-	static public function from(object $source, Locale $locale, array $options = []): LocalizedObject
-	{
-		return new static($source, $locale, $options); // @phpstan-ignore-line
-	}
-
-	/**
-	 * @param T $target The object to localize.
-	 * @param array<string, mixed> $options
+	 * @phpstan-param T $target The object to localize.
 	 */
 	public function __construct(
 		public readonly object $target,
 		public readonly Locale $locale,
-		protected readonly array $options = []
 	) {
 	}
 }